cloud-vm/requirements.txt
google-cloud-bigquery==3.*
scipy==1.10.1
numpy==1.24.3